bugfix updates httpcore/httpclient
git-svn-id: https://svn.code.sf.net/p/tamacat/code/branches/tamacat-httpd-1.4@1175 f3cb667f-c9e3-46e3-bc2c-302c0b7402fb

+++ src/main/java/org/tamacat/httpd/session/FileSessionStore.java
@@ -100,14 +100,15 @@ public class FileSessionStore implements SessionStore {
 
 	@Override
 	public void release() {
-		String[] files = new File(directory).list(
-				getFileNameFilter(fileNameSuffix));
-		for (String f : files) {
-			try {
-				File file = new File(f);
-				file.deleteOnExit();
-			} catch (Exception e) {
-				LOG.warn(e.getMessage());
+		String[] files = new File(directory).list(getFileNameFilter(fileNameSuffix));
+		if (files != null) {
+			for (String f : files) {
+				try {
+					File file = new File(f);
+					file.deleteOnExit();
+				} catch (Exception e) {
+					LOG.warn(e.getMessage());
+				}
 			}
 		}
 	}
